chore: rename to 'collection-route'
