+++ src/app/(main)/collection/page.tsx
@@ -1,5 +1,5 @@
 import { ReactElement } from "react";
-import { OverviewPage } from "@/root/src/libs/client/src/components/modules/collection-page/overview-page";
+import { OverviewPage } from "@/root/src/libs/client/src/components/modules/collection-route/overview-page";
 
 export default function Overview(): ReactElement {
   return <OverviewPage />;
